feat(api): remove api specific code
This removes the workarounds to make the models compatible with Swagger 2.0 for AWS API Gateway.

The task of adapting the model schema to the API definition is
moved out of this repo and will be handled by freyr.

Updates the TypeScript rules to be strict.

BREAKING CHANGE: this changes the underscore prefixes to $